requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Federated Learning
flwr[simulation]>=1.5.0

# Differential Privacy
opacus>=1.4.0

# Model Explainability
shap>=0.42.0

# MLOps and Monitoring
mlflow>=2.8.0
prometheus-client>=0.17.0

# Data Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pathlib2>=2.3.7

# Testing
pytest>=7.4.0
hypothesis>=6.82.0
pytest-cov>=4.1.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Docker and Containerization (for development)
docker>=6.1.0

# Additional utilities
tqdm>=4.65.0
loguru>=0.7.0